Separted bloom functions into their own file, and added Willow license.
