+++ superflore/rosdep_support.py
@@ -1,0 +1,98 @@
+# Software License Agreement (BSD License)
+#
+# Copyright (c) 2013, Willow Garage, Inc.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+#  * Redistributions of source code must retain the above copyright
+#    notice, this list of conditions and the following disclaimer.
+#  * Redistributions in binary form must reproduce the above
+#    copyright notice, this list of conditions and the following
+#    disclaimer in the documentation and/or other materials provided
+#    with the distribution.
+#  * Neither the name of Willow Garage, Inc. nor the names of its
+#    contributors may be used to endorse or promote products derived
+#    from this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+
+from rosdep2 import create_default_installer_context
+from rosdep2.catkin_support import get_catkin_view
+from rosdep2.lookup import ResolutionError
+
+from superflore.exceptions import UnresolvedDependency
+
+import rosdep2.catkin_support
+
+DEFAULT_ROS_DISTRO = 'indigo'
+view_cache = {}
+
+def get_view(os_name, os_version, ros_distro):
+    global view_cache
+    key = os_name + os_version + ros_distro
+    if key not in view_cache:
+        value = get_catkin_view(ros_distro, os_name, os_version, False)
+        view_cache[key] = value
+    return view_cache[key]
+
+
+def resolve_more_for_os(rosdep_key, view, installer, os_name, os_version):
+    """
+    Resolve rosdep key to dependencies and installer key.
+    (This was copied from rosdep2.catkin_support)
+
+    :param os_name: OS name, e.g. 'ubuntu'
+    :returns: resolved key, resolved installer key, and default installer key
+
+    :raises: :exc:`rosdep2.ResolutionError`
+    """
+    d = view.lookup(rosdep_key)
+    ctx = create_default_installer_context()
+    os_installers = ctx.get_os_installer_keys(os_name)
+    default_os_installer = ctx.get_default_os_installer_key(os_name)
+    inst_key, rule = d.get_rule_for_platform(os_name, os_version,
+                                             os_installers,
+                                             default_os_installer)
+    assert inst_key in os_installers
+    return installer.resolve(rule), inst_key, default_os_installer
+
+
+def resolve_rosdep_key(
+    key,
+    os_name,
+    os_version,
+    ros_distro=None,
+    ignored=None,
+    retry=True
+):
+    ignored = ignored or []
+    ctx = create_default_installer_context()
+    try:
+        installer_key = ctx.get_default_os_installer_key(os_name)
+    except KeyError:
+        BloomGenerator.exit("Could not determine the installer for '{0}'"
+                            .format(os_name))
+    installer = ctx.get_installer(installer_key)
+    ros_distro = ros_distro or DEFAULT_ROS_DISTRO
+    view = get_view(os_name, os_version, ros_distro)
+    try:
+        return resolve_more_for_os(key, view, installer, os_name, os_version)
+    except (KeyError, ResolutionError):
+        raise UnresolvedDependency(
+            "could not resolve package {} for os {}."
+            .format(key, os_name)
+        )
